requirements.txt
streamlit
pandas
Pillow
python-dotenv
google-genai
moviepy
edge-tts
